Handle NON_LVALUE_EXPR case in compute_reloc_for_constant.
 A clean of NON_LVALUE_EXPR has be done
(http://www.archivum.info/gcc-patches@gcc.gnu.org/2008-04/01115/PATCH-NON_LVALUE_EXPR-cleanup.html)
but unfornately it seems it's still possible to have a NON_LVALUE_EXPR in compute_reloc_for_constant()
If not handle then me miss reloc possiblity and so variable is put into a ro
segment whereas there is a relocation on it. So we got a relocation into a ro segment which
prevent executable or share library to be load.